+++ src/CoreConfigType.ts
